requirements.txt
uwsgi==2.0.14
psutil==5.1.0
django==1.10.6
django-webpack-loader==0.4.1
djangorestframework==3.6.2
